Auth first cut

+++ src/routes/api/articles/[title]/+server.ts
@@ -1,11 +1,17 @@
 import { mergeIntoGraph, removeFromGraph } from "$lib/api/neo4j";
+import { getUser } from "$lib/api/auth";
 import { deleteArticle, patchAndRenameArticle, patchArticle } from "$lib/api/storage";
 import { extractNames } from "$lib/article/diff";
 import { parse } from "$lib/article/parse";
 import type { RequestHandler } from "./$types";
 import * as db from "$lib/api/db";
+import { error } from "@sveltejs/kit";
 
 export const PATCH: RequestHandler = async (event) => {
+  const user = await getUser(event.cookies);
+  if (user === null) return error(401, "You must be logged in to edit articles.");
+  if (!user.isAdmin) return error(403, "You do not have permission to edit articles.");
+
   const name = event.params.title;
   const patch = await event.request.text();
   const [oldName, newName] = extractNames(patch);
@@ -32,6 +38,10 @@ export const PATCH: RequestHandler = async (event) => {
 };
 
 export const DELETE: RequestHandler = async (event) => {
+  const user = await getUser(event.cookies);
+  if (user === null) return error(401, "You must be logged in to delete articles.");
+  if (!user.isAdmin) return error(403, "You do not have permission to delete articles.");
+
   const title = event.params.title;
   await removeFromGraph(title);
   await deleteArticle(title);
